+++ Library/Homebrew/extend/os/bottle_specification.rb
@@ -1,0 +1,4 @@
+# typed: strict
+# frozen_string_literal: true
+
+require "extend/os/linux/bottle_specification" if OS.linux?
