+++ Library/Homebrew/extend/os/linux/bottle_specification.rb
