--- Library/Homebrew/extend/os/software_spec.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-require "extend/os/linux/software_spec" if OS.linux?
