+++ Library/Homebrew/formula.rb
@@ -19,6 +19,10 @@ require "build_environment"
 require "build_options"
 require "formulary"
 require "software_spec"
+require "bottle"
+require "pour_bottle_check"
+require "head_software_spec"
+require "bottle_specification"
 require "livecheck"
 require "service"
 require "install_renamed"
